refactor: update structure of residuals module
