+++ chemotools/outliers/__init__.py
@@ -1,7 +1,7 @@
-from .dmodx import DModX
-from .hotelling_t2 import HotellingT2
-from .q_residuals import QResiduals
-from .leverage import Leverage
-from .studentized_residuals import StudentizedResiduals
+from ._dmodx import DModX
+from ._hotelling_t2 import HotellingT2
+from ._q_residuals import QResiduals
+from ._leverage import Leverage
+from ._studentized_residuals import StudentizedResiduals
 
 __all__ = ["DModX", "HotellingT2", "QResiduals", "Leverage", "StudentizedResiduals"]
