+++ chemotools/outliers/_hotelling_t2.py
