+++ chemotools/outliers/_q_residuals.py
